move aggregate helpers to monolith package
