+++ custom/icds_reports/utils/monolith/agg_ccs_record.py
