+++ custom/icds_reports/utils/monolith/awc_infrastructure.py
