+++ custom/icds_reports/utils/monolith/awc_location.py
