+++ custom/icds_reports/utils/monolith/ccs_record_monthly.py
