+++ custom/icds_reports/utils/monolith/complementary_forms_ccs_record.py
