+++ custom/icds_reports/utils/monolith/inactive_awws.py
