+++ custom/icds_reports/utils/monolith/thr_froms_ccs_record.py
